refactor: put all tests together
